pyproject.toml
# pyright: reportMissingImports=false
[project]
name = "aiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "pillow>=12.1.0",
    "pip>=26.0.1",
    "python-dotenv==1.1.0",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
    "streamlit-mic-recorder>=0.0.8",
]
